Created pvfactors' subpackage geometry: decoupling from irradiance
