+++ pvfactors/geometry/__init__.py
@@ -1,0 +1,6 @@
+"""Geometry classes"""
+
+from pvfactors.geometry.base import BaseSide, ShadeCollection
+from pvfactors.geometry.pvrow import PVRowSide, PVRow
+from pvfactors.geometry.pvsurface import \
+    BaseSurface, PVSurface, PVSegment
