+++ pvfactors/tests/test_geometry/test_pvrow.py
@@ -1,0 +1,15 @@
+from pvfactors.geometry import PVRow
+from shapely.geometry import LineString
+
+
+def test_pvrow(pvrow_side):
+    """Test that can successfully create a PVRow object from 1 PVRow side,
+    with a shaded pv surface"""
+    pvrow = PVRow(front_side=pvrow_side)
+    assert pvrow.length == 2
+    assert pvrow.front.shaded_length == 1
+    assert pvrow.back.length == 0
+
+    # Check that can find out if intersection
+    line = LineString([(1, 1), (-1, -1)])
+    assert pvrow.intersects(line)
