+++ pvfactors/tests/test_geometry/test_pvsurface.py
@@ -1,5 +1,4 @@
 from pvfactors.pvsurface import PVSegment
-from pvfactors.pvrow import PVRowSide, PVRow
 from shapely.geometry import LineString
 
 
@@ -25,7 +24,7 @@ def test_pvsegment_deleter(shade_collections):
     assert seg.length == 0
 
 
-def test_shaded_length(shade_collections):
+def test_segment_shaded_length(shade_collections):
     """Test that calculation of shaded length is correct"""
     illum_col, shaded_col = shade_collections
     seg_1 = PVSegment(
@@ -34,19 +33,3 @@ def test_shaded_length(shade_collections):
     seg_2 = PVSegment(
         shaded_collection=shaded_col)
     assert seg_2.shaded_length == 1
-
-    side = PVRowSide([seg_1, seg_2])
-    assert side.shaded_length == 1
-
-
-def test_pvrow(pvrow_side):
-    """Test that can successfully create a PVRow object from 1 PVRow side,
-    with a shaded pv surface"""
-    pvrow = PVRow(front_side=pvrow_side)
-    assert pvrow.length == 2
-    assert pvrow.front.shaded_length == 1
-    assert pvrow.back.length == 0
-
-    # Check that can find out if intersection
-    line = LineString([(1, 1), (-1, -1)])
-    assert pvrow.intersects(line)
